% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{premiums_b}
\alias{premiums_b}
\title{Medicare Part B Premiums by Year}
\format{
\subsection{\code{premiums_b}}{

A data frame with 21 rows and 2 columns:
\describe{
\item{year}{Tax year}
\item{premium}{Monthly premium for Medicare Part B}
}
}
}
\source{
\url{https://www.healthline.com/health/medicare/medicare-part-b-premium-increase-history#part-b-increase-history}
}
\usage{
premiums_b
}
\description{
Data set that gives the Medicare Part B monthly premiums by year.
}
\details{
It is a very common misconception that "you get Medicare for free at age 65."
Part A (hospital insurance) is usually premium-free. But Part B (outpatient) always has a monthly
premium.

Most medicare enrollees will pay the standard premium for Medicare Part B. High-income medicare enrollees
have to pay additional premiums due to IRMAA (Income-Related Monthly Adjustment Amount) if income is above
a certain threshold. See \code{irmaa} data for IRMAA thresholds.
}
\keyword{datasets}
